[ticket/10748] Move to namespace and deduplicate array()
PHPBB3-10748
