+++ phpBB/phpbb/db/tools/tools_interface.php
@@ -11,7 +11,7 @@
 *
 */
 
-namespace phpbb\db;
+namespace phpbb\db\tools;
 
 /**
  * Interface for a Database Tools for handling cross-db actions such as altering columns, etc.
